Leeloo/requirements.txt
alembic==1.7.7
aniso8601==8.0.0
apturl==0.5.2
attrs==21.4.0
bidict==0.22.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorzero==1.1
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
dominate==2.6.0
email-validator==1.1.3
entrypoints==0.3
evdev==1.3.0
eventlet==0.33.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-MQTT==1.1.1
Flask-RESTful==0.3.7
flask-restx==0.5.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gevent==21.12.0
gevent-websocket==0.10.1
gpiozero==1.5.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==4.5.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.2.0
MarkupSafe==1.1.1
monotonic==1.5
netifaces==0.10.4
numpy==1.22.4
oauthlib==3.1.0
olefile==0.46
paho-mqtt==1.6.1
pexpect==4.6.0
pigpio==1.44
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-dotenv==0.20.0
python-engineio==3.9.3
python-socketio==4.3.1
python-xlib==0.23
pytz==2019.2
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.4.23
systemd-python==234
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
visitor==0.1.3
wadllib==1.3.3
Werkzeug==0.16.0
WTForms==2.3.3
xkit==0.0.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
